+++ tests/test_process_ei312_cutoff_increment.py
@@ -27,7 +27,8 @@ system_model = "cutoff"
 scenarios = [
     {"model": "remind", "pathway": "SSP3-rollBack", "year": 2050},
     {"model": "image", "pathway": "SSP2-VLHO", "year": 2050},
-    {"model": "tiam-ucl", "pathway": "SSP2-RCP19", "year": 2050},
+    {"model": "tiam-ucl", "pathway": "SSP2-RCP60", "year": 2050},
+    {"model": "message", "pathway": "SSP2-M", "year": 2050},
 ]